+++ src/main/java/pro/sky/demoCollections/service/DepartamentService.java
@@ -2,7 +2,6 @@ package pro.sky.demoCollections.service;
 
 import pro.sky.demoCollections.model.Employee;
 
-import java.math.BigDecimal;
 import java.util.Collection;
 import java.util.List;
 import java.util.Map;
@@ -11,7 +10,7 @@ public interface DepartamentService {
     Employee findMaxSalaryInDepartment(int department);
     Employee findMinSalaryInDepartment(int department);
 
-    BigDecimal printSalaryPerDepartment(int department);
+    Integer printSalaryPerDepartment(int department);
 
     Collection<Employee> printAllFromDepartment(int department);
